hf-rag/requirements.txt
fastapi
uvicorn[standard]
python-multipart

langchain
langchain-google-genai
langchain-text-splitters
langchain-core

google-generativeai

pinecone-client

pydantic-settings
python-dotenv
